Add cache module so it can be reused
