+++ src/version.rs
@@ -1,3 +1,4 @@
+use serde_json::{Value};
 use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize, Debug)]
@@ -10,4 +11,16 @@ pub struct Asset {
 pub struct Version {
   pub tag_name: String,
   pub assets: Vec<Asset>
+}
+
+pub fn get(contents: String) -> Vec<Version> {
+  let raw_versions: Vec<Value> = serde_json::from_str(contents.as_str()).unwrap();
+  let mut versions: Vec<Version> = Vec::new();
+  for raw_version in raw_versions {
+    if raw_version["tag_name"].as_str().unwrap().len() > 0 {
+      let version: Version = serde_json::from_value(raw_version).unwrap();
+      versions.push(version);
+    }
+  }
+  return versions;
 }
